+++ core/autocomplete/filtering/test/filter.test.ts
@@ -1,4 +1,4 @@
-import { setUpTestDir, tearDownTestDir } from "../../../test/util/testDir";
+import { setUpTestDir, tearDownTestDir } from "../../../test/testDir";
 
 import { TEST_CASES_WITH_DIFF, TEST_CASES_WITHOUT_DIFF } from "./testCases";
 import {
