+++ core/config/ConfigHandler.test.ts
@@ -1,8 +1,8 @@
 import fs from "node:fs";
 import path from "node:path";
 
-import { testConfigHandler } from "../test/util/fixtures";
-import { TEST_DIR } from "../test/util/testDir";
+import { testConfigHandler } from "../test/fixtures";
+import { TEST_DIR } from "../test/testDir";
 import { getConfigTsPath } from "../util/paths";
 
 import { defaultConfig } from "./default";
