+++ core/indexing/chunk/ChunkCodebaseIndex.test.ts
@@ -1,6 +1,6 @@
 import { jest } from "@jest/globals";
 
-import { testIde } from "../../test/util/fixtures";
+import { testIde } from "../../test/fixtures";
 import {
   mockFileContents,
   mockFilename,
@@ -8,7 +8,7 @@ import {
   testContinueServerClient,
   updateIndexAndAwaitGenerator,
 } from "../test/indexing";
-import { addToTestDir } from "../../test/util/testDir";
+import { addToTestDir } from "../../test/testDir";
 import { DatabaseConnection, SqliteDb } from "../refreshIndex";
 import { IndexResultType } from "../types";
