+++ core/indexing/test/indexing.ts
@@ -6,8 +6,8 @@ import { ChunkCodebaseIndex } from "../chunk/ChunkCodebaseIndex";
 import { tagToString } from "../refreshIndex";
 import { CodebaseIndex, PathAndCacheKey, RefreshIndexResults } from "../types";
 
-import { testIde } from "../../test/util/fixtures";
-import { addToTestDir, TEST_DIR } from "../../test/util/testDir";
+import { testIde } from "../../test/fixtures";
+import { addToTestDir, TEST_DIR } from "../../test/testDir";
 
 export const mockFilename = "test.py";
 export const mockPathAndCacheKey = {
