+++ core/package.json
@@ -59,7 +59,7 @@
     "comment-json": "^4.2.3",
     "dbinfoz": "^0.14.0",
     "diff": "^7.0.0",
-    "dotenv": "^16.4.7",
+    "dotenv": "^16.4.5",
     "fastest-levenshtein": "^1.0.16",
     "follow-redirects": "^1.15.5",
     "google-auth-library": "^9.14.2",
